% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNMatrix.R
\name{i_dist}
\alias{i_dist}
\title{distance matrix for location i and its neighbors}
\usage{
i_dist(i, neighbor_index, s)
}
\arguments{
\item{i}{The index of point to use}

\item{neighbor_index}{The indices of neighbors}

\item{s}{the locations of neighbors}
}
\description{
distance matrix for location i and its neighbors
}
