% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{nbinom2}
\alias{nbinom2}
\title{Negative binomial family}
\usage{
nbinom2(link = "log")
}
\arguments{
\item{link}{The link, it must be log}
}
\description{
This is the NB2 parameterization where the variance scales quadratically
with the mean.
}
