% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNMatrix.R
\name{NNMatrix}
\alias{NNMatrix}
\title{NNMatrix: A wrapper of spConjNNGP to build Nearest Neighbor matrics ####}
\usage{
NNMatrix(coords, n.neighbors, n.omp.threads = 2, search.type = "tree")
}
\arguments{
\item{coords:}{An n x 2 matrix of the observation coordinates in R^2}

\item{n.neighbors:}{Number of neighbors used in the NNGP.}

\item{n.omp.threads:}{A positive integer indicating the number of threads to use
for SMP parallel processing.}

\item{search.type:}{a quoted keyword that specifies type of nearest neighbor
search algorithm. Supported method key words are: "tree" and "brute"both will
yield the same solution but "tree" should be much faster.}
}
\description{
NNMatrix: A wrapper of spConjNNGP to build Nearest Neighbor matrics ####
}
