% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.glmmfields}
\alias{plot.glmmfields}
\title{Plot predictions from an glmmfields model}
\usage{
\method{plot}{glmmfields}(x, type = c("prediction", "spatial-residual",
  "residual-vs-fitted"), ...)
}
\arguments{
\item{x}{An object returned by \code{\link{glmmfields}}}

\item{type}{Type of plot}

\item{...}{Other arguments passed to \code{\link{predict.glmmfields}}}
}
\description{
Plot predictions from an glmmfields model
}
