% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rrfield}
\alias{plot.rrfield}
\title{Plot predictions from an rrfield model}
\usage{
\method{plot}{rrfield}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{\link{rrfield}}}

\item{...}{Other arguments passed to \code{\link{predict.rrfield}}}
}
\description{
Plot predictions from an rrfield model
}

