% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{student_t}
\alias{student_t}
\alias{half_t}
\title{Student-t and half-t priors}
\usage{
student_t(df = 3, location = 0, scale = 1)

half_t(df = 3, location = 0, scale = 1)
}
\arguments{
\item{df}{Degrees of freedom parameter}

\item{location}{Location parameter}

\item{scale}{Scale parameter}
}
\description{
Student-t and half-t priors. Note that this can be used to represent an
effectively normal distribution prior by setting the first argument (the
degrees of freedom parameter) to a large value (roughly 50 or above).
}
\examples{
student_t(3, 0, 1)
half_t(3, 0, 1)
}
