% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{rrfield}
\alias{rrfield}
\title{Fit a robust spatiotemporal random fields model}
\usage{
rrfield(formula, data, time, lon, lat, nknots = 25L,
  prior_gp_scale = student_t(3, 0, 10), prior_gp_sigma = student_t(3, 0, 2),
  prior_sigma = student_t(3, 0, 2), fixed_df_value = 5,
  estimate_df = TRUE, obs_error = c("normal", "gamma"),
  correlation = c("gaussian", "exponential"), ...)
}
\description{
Fit a robust spatiotemporal random fields model
}

