% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_rrfield}
\alias{sim_rrfield}
\title{Simulate a random field with a MVT distribution}
\usage{
sim_rrfield(n_knots = 15, n_draws = 10, gp_scale = 0.5, gp_sigma = 0.2,
  mvt = TRUE, df = 4, seed = NULL, nDataPoints = 100, sd_obs = 0.1,
  correlation = "gaussian", obs_error = c("normal", "gamma", "nb2"),
  b0 = 0)
}
\arguments{
\item{n_knots}{The number of knots}

\item{n_draws}{The number of draws (for example, the number of years)}

\item{gp_scale}{The Gaussian Process scale parameter}

\item{gp_sigma}{The Gaussian Process sigma parameter}

\item{mvt}{Logical: MVT? (vs. MVN)}

\item{df}{The degrees of freedom parameter for the MVT distribution}

\item{seed}{The random seed value}

\item{nDataPoints}{The number of data points}

\item{sd_obs}{The observation process scale parameter}

\item{correlation}{The correlation or covariance function}

\item{obs_error}{The observation error distribution}

\item{b0}{The intercept parameter}
}
\description{
Simulate a random field with a MVT distribution
}

