% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{stan_pars}
\alias{stan_pars}
\title{Return a vector of parameters}
\usage{
stan_pars(obs_error, estimate_df = TRUE, est_temporalRE = FALSE)
}
\arguments{
\item{obs_error}{The observation error distribution}

\item{estimate_df}{Logical indicating whether the degrees of freedom
parameter should be estimated}
}
\description{
Return a vector of parameters
}

