% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-summary.R
\name{tidy}
\alias{tidy}
\alias{tidy.glmmfields}
\title{Tidy model output}
\usage{
tidy(x, ...)

\method{tidy}{glmmfields}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{glmmfields}}}

\item{...}{Other arguments}
}
\description{
Tidy model output
}
