% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rrfield}
\alias{predict.rrfield}
\title{Predict from a rrfield model}
\usage{
\method{predict}{rrfield}(object, newdata, mcmc_draws, ...)
}
\arguments{
\item{object}{An object returned by \code{\link{rrfield}}.}

\item{newdata}{A data frame to predict on}

\item{mcmc_draws}{The number of MCMC samples to draw from the posterior}

\item{...}{Ignored currently}
}
\description{
Predict from a rrfield model
}

