% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rrfield}
\alias{predict.rrfield}
\title{Predict from an rrfield model}
\usage{
\method{predict}{rrfield}(object, newdata = NULL,
  estimate_method = c("median", "mean"), conf_level = 0.95,
  interval = c("confidence", "prediction"), ...)
}
\arguments{
\item{object}{An object returned by \code{\link{rrfield}}.}

\item{newdata}{Optionally, a data frame to predict on}

\item{estimate_method}{Method for computing point estimate ("mean" or median")}

\item{conf_level}{Probability level for CI}

\item{interval}{Type of interval calculation. Same as \code{\link[stats]{predict.lm}}}

\item{...}{Ignored currently}
}
\description{
Predict from an rrfield model
}

