% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{rrfield}
\alias{rrfield}
\title{Fit a robust spatiotemporal random fields model}
\usage{
rrfield(formula, data, time, lon, lat, station = "", nknots = 25L,
  prior_gp_scale = student_t(3, 0, 5), prior_gp_sigma = student_t(3, 0, 5),
  prior_sigma = student_t(3, 0, 5), prior_rw_sigma = student_t(3, 0, 5),
  prior_intercept = student_t(3, 0, 10), prior_beta = student_t(3, 0, 2),
  fixed_df_value = 5, fixed_ar_value = 0, estimate_df = TRUE,
  estimate_ar = FALSE, obs_error = c("normal", "gamma", "nb2"),
  covariance = c("squared-exponential", "exponential"),
  algorithm = c("sampling", "meanfield"), year_re = FALSE, ...)
}
\arguments{
\item{formula}{The model formula}

\item{data}{A data frame}

\item{time}{A character object giving the name of the time column}

\item{lon}{A character object giving the name of the longitude column}

\item{lat}{A character object giving the name of the latitude column}

\item{station}{A numeric vector giving the integer ID of the station}

\item{nknots}{The number of knots to use in the predictive process model}

\item{prior_gp_scale}{The prior on the Gaussian Process scale parameter. Must
be declared with \code{\link{half_t}}.}

\item{prior_gp_sigma}{The prior on the Gaussian Process sigma parameter. Must
be declared with \code{\link{half_t}}.}

\item{prior_sigma}{The prior on the observation process scale parameter. Must
be declared with \code{\link{half_t}}. This acts as a
substitute for the scale parameter in whatever observation distribution is
being used. I.e. the CV for the Gamma or the dispersion parameter for
the negative binomial.}

\item{prior_rw_sigma}{The prior on the standard deviation parameter of the
random walk process (if specified). Must be declared with
\code{\link{half_t}}.}

\item{prior_intercept}{The prior on the intercept parameter. Must be declared
with \code{\link{student_t}}.}

\item{prior_beta}{The prior on the slope parameters (if any). Must be
declared with \code{\link{student_t}}.}

\item{fixed_df_value}{The fixed value for the student-t degrees of freedom
parameter if the degrees of freedom parameter is fixed. If the degrees of
freedom parameter is estimated then this argument is ignored.}

\item{fixed_ar_value}{The fixed value for temporal autoregressive parameter,
between random fields at time(t) and time(t-1). If the ar parameter
is estimated then this argument is ignored.}

\item{estimate_df}{Logical: should the degrees of freedom perameter be
estimated?}

\item{estimate_ar}{Logical: should the ar parameter be
estimated?}

\item{obs_error}{Character object indicating the observation process
distribution.}

\item{covariance}{Character object describing the covariance
function of the Gaussian Process.}

\item{algorithm}{Character object describing whether the model should be fit
with full NUTS MCMC or via the variational inference mean-field approach.
See \code{\link[rstan]{vb}}. Note that the variational inference approach
should not be trusted for final inference and is much more likely to give
incorrect inference than MCMC.}

\item{year_re}{Logical: estimate a random walk for the time variable? If
\code{TRUE}, then no fixed effects (B coefficients) will be estimated.}

\item{...}{Any other arguments to pass to \code{\link[rstan]{sampling}}.}
}
\description{
Fit a robust spatiotemporal random fields model
}

