% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_rrfield}
\alias{sim_rrfield}
\title{Simulate a random field with a MVT distribution}
\usage{
sim_rrfield(n_knots = 15, n_draws = 10, gp_scale = 0.5, gp_sigma = 0.2,
  mvt = TRUE, df = 4, seed = NULL, n_data_points = 100, sd_obs = 0.1,
  covariance = "squared-exponential", obs_error = c("normal", "gamma",
  "poisson", "nb2", "binomial", "lognormal"), B = c(0), ar = 0, X = rep(1,
  n_draws * n_data_points), demean = FALSE)
}
\arguments{
\item{n_knots}{The number of knots}

\item{n_draws}{The number of draws (for example, the number of years)}

\item{gp_scale}{The Gaussian Process scale parameter}

\item{gp_sigma}{The Gaussian Process sigma parameter}

\item{mvt}{Logical: MVT? (vs. MVN)}

\item{df}{The degrees of freedom parameter for the MVT distribution}

\item{seed}{The random seed value}

\item{n_data_points}{The number of data points per draw}

\item{sd_obs}{The observation process scale parameter}

\item{covariance}{The covariance function}

\item{obs_error}{The observation error distribution}

\item{B}{A vector of parameters. The first element is the intercept}

\item{ar}{The auto regressive parameter on the mean of the random field knots}

\item{X}{The model matrix}

\item{demean}{Should the spatial knots be centered by subtracting
their mean at each time slice?}
}
\description{
Simulate a random field with a MVT distribution
}
